requirements.txt
# Core dependencies
requests>=2.28.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.3
pytest-cov>=4.0.0  # For coverage reporting

# Additional dependencies
aiohttp>=3.9.1
python-dotenv>=1.0.0
typing-extensions>=4.8.0 